fix: 모달 오픈시 화면이 떨리는 현상을 해결한다
