+++ src/utils/fixScrollbar.ts
@@ -1,0 +1,20 @@
+import { isServer } from './ssr';
+
+export function fixScrollbar() {
+  if (isServer()) {
+    return 0;
+  }
+  const outer = document.createElement('div');
+  outer.style.visibility = 'hidden';
+  outer.style.overflow = 'scroll';
+  outer.style.msOverflowStyle = 'scrollbar';
+  document.body.appendChild(outer);
+
+  const inner = document.createElement('div');
+  outer.appendChild(inner);
+
+  const scrollbarWidth = outer.offsetWidth - inner.offsetWidth;
+
+  outer.parentNode && outer.parentNode.removeChild(outer);
+  return scrollbarWidth;
+}
