+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './ifDarkTheme';
 export * from './defaultKeyExtractor';
 export * from './memoComponent';
 export * from './ssr';
+export * from './fixScrollbar';
